+++ src/main/settings.py
@@ -1,4 +1,4 @@
-from framework.util.settings import get_setting
+from framework.utils.settings import get_setting
 
 HOST = get_setting("HOST", "localhost")
 PORT = get_setting("PORT", 8000, convert=int)
