+++ src/scripts/get_db_name.py
@@ -1,4 +1,4 @@
-from framework.util.db import get_db_name
+from framework.utils.db import get_db_name
 
 
 def main():
